+++ CHANGELOG.md
@@ -13,6 +13,21 @@
   * Many instances of database entities being dropped and recreated instead of altered are now supported.
 * [#47](https://github.com/obsidiansystems/beam-automigrate/pull/47): Generate postgres enum types in schema when using `Nullable PgEnum` values.
 
+## 0.1.6.0
+
+* Fix instance HasSchemaConstraints build failure with GHC 9.2.8
+
+## 0.1.5.0
+
+* Add ltree column type
+* Add vector column type
+* Fix ghc 9.2.8 build
+
+## 0.1.4.0
+
+* [#52](https://github.com/obsidiansystems/beam-automigrate/pull/52) Support sql arrays
+* Loosen some version bounds
+
 ## 0.1.3.0
 
 * [#47](https://github.com/obsidiansystems/beam-automigrate/pull/47): Generate postgres enum types in schema when using `Nullable PgEnum` values.